requirements.txt
html-sanitizer
nicegui
python-dotenv